Obtain List of Jacobian Matrix Entries for Covariance Matrix

h_jac_list(covbeta_fun, theta_est)

Arguments

covbeta_fun

(function)
function calculating the covariance matrix of coefficients given variance parameters (theta), see h_covbeta_fun() to obtain this from a mmrm_tmb object.

theta_est

(numeric)
variance parameters point estimate.

Value

List with one element per variance parameter containing a matrix of the same dimensions as the covariance matrix. The values are the derivatives with regards to this variance parameter.

Examples

formula <- FEV1 ~ RACE + us(AVISIT | USUBJID)
model <- h_mmrm_tmb(formula, fev_data)
covbeta_fun <- h_covbeta_fun(model)
h_jac_list(covbeta_fun, model$theta_est)
#> [[1]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.02703863 -0.02703863 -0.02703863
#> [2,] -0.02703863  0.06933062  0.02703863
#> [3,] -0.02703863  0.02703863  0.06974491
#> 
#> [[2]]
#>            [,1]       [,2]       [,3]
#> [1,]  0.1384094 -0.1384094 -0.1384094
#> [2,] -0.1384094  0.2903041  0.1384094
#> [3,] -0.1384094  0.1384094  0.3269200
#> 
#> [[3]]
#>           [,1]       [,2]      [,3]
#> [1,]  0.271299 -0.2712990 -0.271299
#> [2,] -0.271299  0.5279481  0.271299
#> [3,] -0.271299  0.2712990  0.645867
#> 
#> [[4]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.03549032 -0.03549032 -0.03549032
#> [2,] -0.03549032  0.09156623  0.03549032
#> [3,] -0.03549032  0.03549032  0.09246775
#> 
#> [[5]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.05876961 -0.05876961 -0.05876961
#> [2,] -0.05876961  0.12273789  0.05876961
#> [3,] -0.05876961  0.05876961  0.13608450
#> 
#> [[6]]
#>            [,1]       [,2]       [,3]
#> [1,]  0.1000752 -0.1000752 -0.1000752
#> [2,] -0.1000752  0.1874225  0.1000752
#> [3,] -0.1000752  0.1000752  0.2209471
#> 
#> [[7]]
#>            [,1]       [,2]       [,3]
#> [1,]  0.1292216 -0.1292216 -0.1292216
#> [2,] -0.1292216  0.2400591  0.1292216
#> [3,] -0.1292216  0.1292216  0.2981726
#> 
#> [[8]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.03186863 -0.03186863 -0.03186863
#> [2,] -0.03186863  0.08519624  0.03186863
#> [3,] -0.03186863  0.03186863  0.08173603
#> 
#> [[9]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.04179025 -0.04179025 -0.04179025
#> [2,] -0.04179025  0.10050280  0.04179025
#> [3,] -0.04179025  0.04179025  0.09797409
#> 
#> [[10]]
#>             [,1]        [,2]        [,3]
#> [1,]  0.05258082 -0.05258082 -0.05258082
#> [2,] -0.05258082  0.11415496  0.05258082
#> [3,] -0.05258082  0.05258082  0.13674006
#>