Methods for mmrm_tmb Objects

# S3 method for mmrm_tmb
coef(object, ...)

# S3 method for mmrm_tmb
logLik(object, ...)

# S3 method for mmrm_tmb
formula(x, ...)

# S3 method for mmrm_tmb
vcov(object, ...)

# S3 method for mmrm_tmb
VarCorr(x, sigma = NA, ...)

# S3 method for mmrm_tmb
deviance(object, ...)

# S3 method for mmrm_tmb
AIC(object, corrected = FALSE, ..., k = 2)

# S3 method for mmrm_tmb
BIC(object, ...)

Arguments

object

(mmrm_tmb)
the fitted MMRM object.

...

not used.

x

(mmrm_tmb)
same as object.

sigma

cannot be used (this parameter does not exist in MMRM).

corrected

(flag)
whether corrected AIC should be calculated.

k

(number)
the penalty per parameter to be used; default k = 2 is the classical AIC.

Functions

coef.mmrm_tmb: obtains the estimated coefficients.

logLik.mmrm_tmb: obtains the attained log likelihood value.

formula.mmrm_tmb: obtains the used formula.

vcov.mmrm_tmb: obtains the variance-covariance matrix estimate for the coefficients.

VarCorr.mmrm_tmb: obtains the variance-covariance matrix estimate for the residuals.

deviance.mmrm_tmb: obtains the deviance, which is defined here as twice the negative log likelihood, which can either be integrated over the coefficients for REML fits or the usual one for ML fits.

AIC.mmrm_tmb: obtains the Akaike Information Criterion, where the degrees of freedom are the number of variance parameters (n_theta). If corrected, then this is multiplied with m / (m - n_theta - 1) where m is the number of observations minus the number of coefficients, or n_theta + 2 if it is smaller than that.

BIC.mmrm_tmb: obtains the Bayesian Information Criterion, which is using the natural logarithm of the number of subjects for the penalty parameter k.

Examples

formula <- FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
object <- h_mmrm_tmb(formula, fev_data)
# Estimated coefficients:
coef(object)
#>                   (Intercept) RACEBlack or African American 
#>                    30.7774010                     1.5305918 
#>                     RACEWhite                     SEXFemale 
#>                     5.6435688                     0.3260376 
#>                      ARMCDTRT                    AVISITVIS2 
#>                     3.7744170                     4.8396061 
#>                    AVISITVIS3                    AVISITVIS4 
#>                    10.3421689                    15.0537862 
#>           ARMCDTRT:AVISITVIS2           ARMCDTRT:AVISITVIS3 
#>                    -0.0420969                    -0.6938117 
#>           ARMCDTRT:AVISITVIS4 
#>                     0.6241162 
# Log likelihood given the estimated parameters:
logLik(object)
#> [1] -1693.225
# Formula which was used:
formula(object)
#> FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID)
#> <environment: 0x558b0adc2e38>
# Variance-covariance matrix estimate for coefficients:
vcov(object)
#>                               (Intercept) RACEBlack or African American
#> (Intercept)                     0.7859969                  -0.226202584
#> RACEBlack or African American  -0.2262026                   0.389949546
#> RACEWhite                      -0.1771022                   0.181456474
#> SEXFemale                      -0.1684043                   0.031536614
#> ARMCDTRT                       -0.5674920                   0.028373173
#> AVISITVIS2                     -0.4227692                   0.002975203
#> AVISITVIS3                     -0.5231225                   0.010827226
#> AVISITVIS4                     -0.4407389                   0.002206353
#> ARMCDTRT:AVISITVIS2             0.4225399                   0.005381310
#> ARMCDTRT:AVISITVIS3             0.5218961                   0.011420293
#> ARMCDTRT:AVISITVIS4             0.4490212                  -0.012591464
#>                                  RACEWhite    SEXFemale     ARMCDTRT
#> (Intercept)                   -0.177102161 -0.168404273 -0.567492041
#> RACEBlack or African American  0.181456474  0.031536614  0.028373173
#> RACEWhite                      0.443009672  0.023362344 -0.042966923
#> SEXFemale                      0.023362344  0.282955706  0.001813942
#> ARMCDTRT                      -0.042966923  0.001813942  1.153814377
#> AVISITVIS2                    -0.003147370  0.006471253  0.419540240
#> AVISITVIS3                    -0.002951286  0.006770353  0.517277178
#> AVISITVIS4                    -0.008230918  0.004088502  0.440748366
#> ARMCDTRT:AVISITVIS2            0.013485997 -0.016800949 -0.845781723
#> ARMCDTRT:AVISITVIS3            0.006720665 -0.024695337 -1.044354125
#> ARMCDTRT:AVISITVIS4            0.002966413 -0.009039702 -0.877812748
#>                                 AVISITVIS2   AVISITVIS3   AVISITVIS4
#> (Intercept)                   -0.422769225 -0.523122498 -0.440738922
#> RACEBlack or African American  0.002975203  0.010827226  0.002206353
#> RACEWhite                     -0.003147370 -0.002951286 -0.008230918
#> SEXFemale                      0.006471253  0.006770353  0.004088502
#> ARMCDTRT                       0.419540240  0.517277178  0.440748366
#> AVISITVIS2                     0.642763903  0.399030619  0.368462962
#> AVISITVIS3                     0.399030619  0.676812701  0.401841357
#> AVISITVIS4                     0.368462962  0.401841357  1.723666955
#> ARMCDTRT:AVISITVIS2           -0.643034241 -0.399184844 -0.368746904
#> ARMCDTRT:AVISITVIS3           -0.399220399 -0.676473418 -0.401834200
#> ARMCDTRT:AVISITVIS4           -0.368629535 -0.402209189 -1.723775072
#>                               ARMCDTRT:AVISITVIS2 ARMCDTRT:AVISITVIS3
#> (Intercept)                            0.42253988         0.521896096
#> RACEBlack or African American          0.00538131         0.011420293
#> RACEWhite                              0.01348600         0.006720665
#> SEXFemale                             -0.01680095        -0.024695337
#> ARMCDTRT                              -0.84578172        -1.044354125
#> AVISITVIS2                            -0.64303424        -0.399220399
#> AVISITVIS3                            -0.39918484        -0.676473418
#> AVISITVIS4                            -0.36874690        -0.401834200
#> ARMCDTRT:AVISITVIS2                    1.27538728         0.805808529
#> ARMCDTRT:AVISITVIS3                    0.80580853         1.410477398
#> ARMCDTRT:AVISITVIS4                    0.72897499         0.796509353
#>                               ARMCDTRT:AVISITVIS4
#> (Intercept)                           0.449021184
#> RACEBlack or African American        -0.012591464
#> RACEWhite                             0.002966413
#> SEXFemale                            -0.009039702
#> ARMCDTRT                             -0.877812748
#> AVISITVIS2                           -0.368629535
#> AVISITVIS3                           -0.402209189
#> AVISITVIS4                           -1.723775072
#> ARMCDTRT:AVISITVIS2                   0.728974993
#> ARMCDTRT:AVISITVIS3                   0.796509353
#> ARMCDTRT:AVISITVIS4                   3.426059691
# Variance-covariance matrix estimate for residuals:
VarCorr(object)
#>           VIS1      VIS2       VIS3       VIS4
#> VIS1 40.554180 14.396017  4.9759858 13.3775460
#> VIS2 14.396017 26.571631  2.7836356  7.4773142
#> VIS3  4.975986  2.783636 14.8979316  0.9035493
#> VIS4 13.377546  7.477314  0.9035493 95.5564606
# REML criterion (twice the negative log likelihood):
deviance(object)
#> [1] 3386.45
# AIC:
AIC(object)
#> [1] 3406.45
AIC(object, corrected = TRUE)
#> [1] 3406.877
# BIC:
BIC(object)
#> [1] 3439.282