Fit an MMRM

This is the main function fitting the MMRM.

Usage

mmrm(
  formula,
  data,
  reml = TRUE,
  optimizer = "automatic",
  n_cores = h_free_cores(),
  accept_singular = TRUE
)

Arguments

formula

(formula)
the model formula, see details.

data

(data)
the data to be used for the model.

reml

(flag)
whether restricted maximum likelihood (REML) estimation is used, otherwise maximum likelihood (ML) is used.

optimizer

(string)
optimizer to be used to generate the model.

n_cores

(count)
number of cores which could in principle be used for parallel computations on Linux or Mac machines.

accept_singular

(flag)
whether singular design matrices are reduced to full rank automatically and additional coefficient estimates will be missing.

Value

An mmrm object.

Details

The formula typically looks like: FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID) so specifies response and covariates as usual, and exactly one special term defines which covariance structure is used and what are the visit and subject variables.

When setting optimizer = "automatic", first the default optimizer (L-BFGS-B) is used to fit the model. If that converges, this is returned. If not, the other available optimizers from refit_multiple_optimizers() are tried (in parallel if n_cores is set and not on Windows). If none of the optimizers converge, then the function fails. Otherwise the best fit is returned.

Note

The mmrm object is also an mmrm_fit and an mmrm_tmb object, therefore corresponding methods also work (see mmrm_tmb_methods). In addition it contains the Jacobian information jac_list and the call. Use of the package emmeans is supported, see emmeans_support.

Examples

fit <- mmrm(
  formula = FEV1 ~ RACE + SEX + ARMCD * AVISIT + us(AVISIT | USUBJID),
  data = fev_data
)